fair-gig-monorepo/services/anomaly-detection-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
httpx==0.23.3
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
motor==3.1.1
dnspython==2.3.0
